# apple-pi/services/startup-core/src/server.ts
// Startup Core Service - Produktübergabe und Verifizierung
// BRANDING: .T. TogetherSystems - ModularFlux Architecture
// VERSION: 3.0.0

import express from 'express';
import cors from 'cors';
import Ajv from 'ajv';
import fs from 'fs';
import path from 'path';

const ajv = new Ajv({ allErrors: true });
const schema = JSON.parse(
  fs.readFileSync(path.join(__dirname, '../../../specs/domain/PP_product.schema.json'), 'utf8')
);
const validate = ajv.compile(schema);

const app = express();
app.use(cors());
app.use(express.json());

// Produkte abrufen
app.get('/startup/products', (_req, res) => {
  // In Produktion: aus DB laden
  res.json([]);
});

// Neues Produkt erstellen
app.post('/startup/products', (req, res) => {
  if (!validate(req.body)) {
    return res.status(400).json({ errors: validate.errors });
  }
  
  const product = {
    ...req.body,
    id: `PP-${Date.now().toString(36).toUpperCase()}`,
    createdAt: new Date().toISOString(),
    status: 'development'
  };
  
  // In Produktion: in DB speichern
  res.status(201).json(product);
});

// Produkt verifizieren
app.post('/startup/products/:productId/verify', async (req, res) => {
  const { productId } = req.params;
  
  // 1. Produkt laden
  // 2. Manifest erstellen
  // 3. Digitalnotator aufrufen
  // 4. Verifizierung speichern
  
  const verification = {
    productId,
    verified: true,
    verifiedAt: new Date().toISOString(),
    notarId: `NN-${Date.now().toString(36)}`,
    hessenNumber: null
  };
  
  res.json(verification);
});

// Entwicklungsberichte
app.post('/startup/products/:productId/reports', (req, res) => {
  const { productId } = req.params;
  const report = req.body;
  
  // Speichere Entwicklungsbericht
  res.status(201).json({ productId, report });
});

// Mathematische Verifikation
app.post('/startup/products/:productId/verify-math', (req, res) => {
  const { productId } = req.params;
  const { calculations } = req.body;
  
  // Prüfe mathematische Korrektheit
  const results = calculations.map(calc => {
    // Vereinfachte Prüfung
    return {
      formula: calc.formula,
      result: eval(calc.formula), // In Produktion: sicherer Parser
      valid: true
    };
  });
  
  res.json({ productId, results });
});

const PORT = process.env.PORT || 8086;
app.listen(PORT, () => {
  console.log(`✅ Startup Core läuft auf Port ${PORT}`);
});

